backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
requests==2.31.0
openai==1.3.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
prophet==1.1.4
python-multipart==0.0.6
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
xgboost==2.0.3
lightgbm==4.1.0
tensorflow==2.16.1
joblib==1.3.2
aiohttp==3.9.3
websockets==12.0
